requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-extensions==1.5.1
django-postgrespool==0.3.0
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
six==1.10.0
SQLAlchemy==1.0.9
static3==0.6.1
urllib3==1.6
virtualenv==13.1.2
django-crispy-forms==1.4.0
django-registration-redux==1.1
oauthlib==1.0.3
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0


